test(auth): cover environment ownership gate
Exercise an unscoped environment route so account-only key denial depends directly on the ownership middleware.
